to byte array tests
